[423562] Apply the same JDT preferences to all plug-ins
This includes the Sirius formatting and cleanup profiles, and fixes all
unnecessary casts, which are now configured as errors.

All settings are "inlined" as project-specific settings to ensure the
behavior is the same for everyone independently of workspace-specific
configuration.

+++ plugins/org.eclipse.sirius.diagram/src/org/eclipse/sirius/diagram/ui/tools/internal/figure/NotificationFigure.java
@@ -101,7 +101,7 @@ public class NotificationFigure extends RectangleFigure {
 
         // Workaround, as the size of the figure follows zoom, we apply twice
         // the zoom on figure offset to balance
-        double notationFigureOffset = ((double) (viewSize.width - WIDTH) / 2d) / (rootEditPart.getZoomManager().getZoom() * rootEditPart.getZoomManager().getZoom());
+        double notationFigureOffset = ((viewSize.width - WIDTH) / 2d) / (rootEditPart.getZoomManager().getZoom() * rootEditPart.getZoomManager().getZoom());
 
         viewLocation.performScale(1.0d / rootEditPart.getZoomManager().getZoom());
